refactor router, daoChainId state out of HausContext.
