+++ libs/daohaus-connect-feature/src/HausConnectContext.tsx
@@ -48,7 +48,7 @@ export type UserConnectType = {
   networks: NetworkConfigs;
   switchNetwork: (chainId: string) => void;
   isProfileLoading: boolean;
-  daoChainId: string | null;
+  daoChainId?: string;
   validNetwork: boolean;
 };
 
@@ -60,6 +60,7 @@ type ConnectProviderProps = {
   networks?: NetworkConfigs;
   children: ReactNode;
   handleModalEvents?: ModalEvents;
+  daoChainId?: string;
 };
 
 export const HausConnectProvider = ({
@@ -67,6 +68,7 @@ export const HausConnectProvider = ({
   children,
   networks = supportedNetworks,
   handleModalEvents,
+  daoChainId,
 }: ConnectProviderProps) => {
   const [isConnecting, setConnecting] = useState(true);
   const [{ provider, chainId, address }, setWalletState] =
@@ -76,7 +78,6 @@ export const HausConnectProvider = ({
     ens: undefined,
   });
   const [isProfileLoading, setProfileLoading] = useState(false);
-  const [daoChainId, setDaoChainId] = useState<string | null>(null);
 
   const isConnected = useMemo(
     () => !!provider && !!address && !!chainId,
@@ -88,9 +89,6 @@ export const HausConnectProvider = ({
     [chainId, networks]
   );
 
-  const location = useLocation();
-  const pathMatch = matchPath('molochv3/:daochain/:daoid/*', location.pathname);
-
   const connectWallet = useCallback(async () => {
     handleConnectWallet({
       setConnecting,
@@ -100,12 +98,6 @@ export const HausConnectProvider = ({
       web3modalOptions,
     });
   }, [setConnecting, handleModalEvents, web3modalOptions]);
-
-  useEffect(() => {
-    if (pathMatch?.params?.daochain) {
-      setDaoChainId(pathMatch?.params?.daochain);
-    }
-  }, [pathMatch?.params?.daochain, setDaoChainId]);
 
   useEffect(() => {
     loadWallet({ setConnecting, connectWallet, web3modalOptions });
